Import live.2010.12.31.tar.gz
- We now support 'trick play' operations (seeking, fast-forward, reverse play) on MPEG Transport Stream files that contain
  H.264 video (rather than just MPEG-2 video, as previously).  To support this, the index file format has been extended in a
  backwards-compatible way, so that existing index files (for MPEG-2 video Transport Streams) will continue to work as before.
  New versions of the "MPEG2TransportStreamIndexer" and "testMPEG2TransportStreamTrickPlay" utilities - and the
  "live555MediaServer" - have also been released.
- Fixed a bug in the definition of the "profile_level_id" field in "H264VideoRTPSink".  (Thanks to Geoff Cleary for noting this.)
- Change the parsing of RTSP "Range:" headers to allow parameters of the form "clock=" or"smtpe=".  (However, we currently don't
  interpret parameters of this form; instead, we just ignore them.)
  (Thanks to Sebastien Escudier for this suggestion.)
- Fixed "openRTSP" to properly reset internal state before repeating the playing of a stream (if the '-c' (play continuously))
  option is used.  (Thanks to Anon Sricharoenchai for noting this.)

+++ liveMedia/H264VideoRTPSink.cpp
@@ -14,7 +14,7 @@ along with this library; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301  USA
 **********/
 // "liveMedia"
-// Copyright (c) 1996-2010 Live Networks, Inc.  All rights reserved.
+// Copyright (c) 1996-2011 Live Networks, Inc.  All rights reserved.
 // RTP sink for H.264 video (RFC 3984)
 // Implementation
 
@@ -112,7 +112,7 @@ char const* H264VideoRTPSink::auxSDPLine() {
   framerSource->getSPSandPPS(sps, spsSize, pps, ppsSize);
   if (sps == NULL || pps == NULL) return NULL; // our source isn't ready
 
-  u_int8_t profile_level_id;
+  u_int32_t profile_level_id;
   if (spsSize < 4) { // sanity check
     profile_level_id = 0;
   } else {
